requirements.txt
# Core RAG stack (LangChain 0.1.14 compatible)
langchain==0.1.14
langchain-core==0.1.53
langchain-openai==0.1.6
langchain-community==0.0.32
langchain-text-splitters==0.0.2
langsmith==0.1.147
pydantic==2.7.1

# Vector DB
chromadb==0.4.24
tiktoken==0.9.0

# App
streamlit>=1.33
python-dotenv>=1.0

# Google APIs (Calendar & Sheets)
google-api-python-client
gspread
google-auth

# Utilities
requests>=2.31



